+++ lib/generators/install/templates/create_permissions.rb
@@ -1,6 +1,6 @@
 class CreatePermissions < ActiveRecord::Migration[<%= ActiveRecord::Migration.current_version %>]
   def change
-    create_table :permissions do |t|
+    create_table :permissions<%= ", id: :uuid" if primary_key_type == :uuid %> do |t|
       t.string :name, null: false
 
       t.timestamps